add everything

+++ detoks/src/main/java/nl/tudelft/trustchain/detoks/BenchmarkResult.kt
@@ -2,9 +2,9 @@ package nl.tudelft.trustchain.detoks
 
 import com.github.mikephil.charting.data.Entry
 
-data class BenchmarkResult (
-    var timePerBlock : ArrayList<Entry>,
-    var totalTime : Long,
-    var payloadBandwith : Double,
-    var lostPackets: Long
+data class BenchmarkResult(
+    var timePerBlock: ArrayList<Entry>,
+    var totalTime: Long,
+    var payloadBandwith: Double,
+    var lostPackets: Int
     )
